moved molbe RDM methods into the molbe class
